Fix NextAuth cant build
