+++ app/admin/UserManagement/edit/[id]/page.tsx
@@ -1,0 +1,3 @@
+export default function UserManagement() {
+  return <div>UserManagement</div>;
+}
