+++ app/api/auth/[...nextauth]/route.ts
@@ -1,90 +1,5 @@
-import { PrismaClient } from "@prisma/client";
-import { PrismaAdapter } from "@next-auth/prisma-adapter";
-import bcrypt from "bcrypt";
-
-import NextAuth, { NextAuthOptions } from "next-auth";
-import CredentialsProvider from "next-auth/providers/credentials";
-import GoogleProvider from "next-auth/providers/google";
-
-const prisma = new PrismaClient();
-
-const clientId = process.env.GOOGLE_CLIENT_ID || "";
-const clientSecret = process.env.GOOGLE_CLIENT_SECRET || "";
-
-if (!clientId || !clientSecret) {
-  throw new Error("Missing Google OAuth environment variables");
-}
-
-export const authOptions: NextAuthOptions = {
-  session: {
-    strategy: "jwt",
-  },
-  secret: process.env.NEXT_AUTH_SECRET,
-  adapter: PrismaAdapter(prisma),
-
-  providers: [
-    CredentialsProvider({
-      name: "Credentials",
-      credentials: {
-        email: { label: "Email", type: "email", placeholder: "Email" },
-        password: { label: "Password", type: "password" },
-      },
-      async authorize(credentials) {
-        if (!credentials?.email || !credentials?.password) {
-          return null;
-        }
-
-        const user = await prisma.user.findUnique({
-          where: {
-            email: credentials.email,
-          },
-        });
-
-        if (
-          user &&
-          user.password &&
-          (await bcrypt.compare(credentials.password, user.password))
-        ) {
-          return {
-            id: user.id + "",
-            name: user.name,
-            email: user.email,
-            role: user.role,
-          };
-        } else {
-          throw new Error("Invalid email or password");
-        }
-      },
-    }),
-
-    GoogleProvider({
-      clientId,
-      clientSecret,
-    }),
-  ],
-
-  callbacks: {
-    jwt: async ({ token, user }) => {
-      if (user) {
-        token.id = user.id;
-        token.role = user.role;
-      }
-      return token;
-    },
-
-    session: async ({ session, token }) => {
-      if (token) {
-        session.user.id = token.id;
-        session.user.role = token.role;
-      }
-      return session;
-    },
-  },
-
-  pages: {
-    signIn: "/admin",
-  },
-};
+import { authOptions } from "@/app/api/auth/[...nextauth]/authOptions";
+import NextAuth from "next-auth";
 
 const handler = NextAuth(authOptions);
 export { handler as GET, handler as POST };
